requirements.txt
pandas
openpyxl
PyYAML
xlrd
dash
dash-bootstrap-components
plotly
highcharts-core
highcharts-maps
